同步到6.83

+++ src/main/java/com/github/jspxnet/txweb/turnpage/TurnPageButton.java
@@ -46,7 +46,7 @@ public interface TurnPageButton {
 
     void setTotalPage(int totalPage);
 
-    int getTotalPage();
+    long getTotalPage();
 
 
     void setCurrentPath(String currentPath);
